requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
twilio==8.10.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3
gunicorn==21.2.0